requirements.txt
flask>=3.0
flask-login>=0.6
flask-caching>=2.0
flask-limiter>=3.5
sqlalchemy>=2.0
psycopg2-binary>=2.9
bcrypt>=4.0
python-dotenv>=1.0
PyJWT>=2.8
werkzeug>=3.0
requests>=2.31
